Event開始/終了時間による動作制御。deleteを削除フラグにて対応。

+++ app/View/Events/detail.ctp
@@ -1,49 +1,55 @@
 <?php echo $this->Html->css('event'); ?>
 
+<?php if($event['Event']['del_flg'] == 1){ throw new NotFoundException(__('Invalid post')); } ?>
+
 <?php echo $this->Upload->uploadImage($event, 'Event.img', array('style' => 'thumb')) ?>
 <h2><?php echo h($event['Event']['title']); ?></h2>
 <?php
-	echo $this->Form->postLink(
-		        'Join',
-	            array('action' => 'join', $event['Event']['id'])
-	            );
-	echo '　　';
-	echo $this->Form->postLink(
-		        'Maybe',
-	            array('action' => 'maybe', $event['Event']['id'])
-	            );
-	echo '　　';
-	echo $this->Form->postLink(
-		        'Decline',
-	            array('action' => 'decline', $event['Event']['id'])
-	            );
-	echo '<br /><br />';
+	if(strtotime($event['Event']['open_datetime']) > strtotime(date("Y-m-d H:i:s"))){
+		echo $this->Form->postLink(
+			        'Join',
+		            array('action' => 'join', $event['Event']['id'])
+		            );
+		echo '　　';
+		echo $this->Form->postLink(
+			        'Maybe',
+		            array('action' => 'maybe', $event['Event']['id'])
+		            );
+		echo '　　';
+		echo $this->Form->postLink(
+			        'Decline',
+		            array('action' => 'decline', $event['Event']['id'])
+		            );
+		echo '<br /><br />';
 
-	//当該イベントへの当該ユーザーのstatusを反映させる
-	if(isset($participants[0]['Participant']['status'])){
-		$participant_status = $participants[0]['Participant']['status'];
+		//当該イベントへの当該ユーザーのstatusを反映させる
+		if(isset($participants[0]['Participant']['status'])){
+			$participant_status = $participants[0]['Participant']['status'];
 
-		if($participant_status == 1){
-			echo '<h3>'.'　　▶︎ You\'re invited this event!'.'</h3>';
-		}else if($participant_status == 2){
-			echo '<h3>'.'　　▶︎ Your status is "Join"'.'</h3>';
-		}else if($participant_status == 3){
-			echo '<h3>'.'　　▶ ︎Your status is "Maybe"'.'</h3>';
-		}else if($participant_status == 4){
-			echo '<h3>'.'　　▶︎ Your status is "Decline"'.'</h3>';
+			if($participant_status == 1){
+				echo '<h3>'.'　　▶︎ You\'re invited this event!'.'</h3>';
+			}else if($participant_status == 2){
+				echo '<h3>'.'　　▶︎ Your status is "Join"'.'</h3>';
+			}else if($participant_status == 3){
+				echo '<h3>'.'　　▶ ︎Your status is "Maybe"'.'</h3>';
+			}else if($participant_status == 4){
+				echo '<h3>'.'　　▶︎ Your status is "Decline"'.'</h3>';
+			}else{
+				echo '<h3>'.'　　Something is wrong'.'</h3>';	//これはべつになくてもいいかも
+			}
 		}else{
-			echo '<h3>'.'　　Something is wrong'.'</h3>';	//これはべつになくてもいいかも
+			echo '<h3>'.'　　▶︎ Please express your will'.'</h3>';
 		}
 	}else{
-		echo '<h3>'.'　　▶︎ Please express your will'.'</h3>';
+		echo '<h3>'.'　　This event has been already finished'.'</h3>';
 	}
 ?>
 
 <hr><br />
 <dt>Date：</dt>
-	<dd><?php echo h(substr($event['Event']['open_datetime'],0,16)); ?>〜<?php echo h(substr($event['Event']['close_datetime'],0,16)); ?></dd>
+	<dd><?php echo date('M.d.Y  H:m', strtotime($event['Event']['open_datetime'])).'  〜  '.date('M.d.Y  H:m', strtotime($event['Event']['close_datetime'])); ?></dd>
 <dt>Hosting：</dt>
-<dd><?php echo h($hosting['User']['name']); ?></dd>
+<dd><?php echo $this->Html->link(h($hosting['User']['name']), array('controller' => 'users', 'action' => 'view', $hosting['User']['id'])); ?></dd>
 <dt>Place：</dt>
 	<dd><?php echo h($event['Event']['place']); ?></dd>
 <dt>Detail：</dt>
@@ -53,7 +59,8 @@
 	<dd>
 		<?php
 			foreach($join_info as $join){
-				echo h($join['User']['name']).'　';
+				echo $this->Html->link(h($join['User']['name']), array('controller' => 'users', 'action' => 'view', $join['User']['id']));
+				echo '　';
 			}
 		?>
 	</dd>
@@ -61,7 +68,8 @@
 	<dd>
 		<?php
 			foreach($maybe_info as $maybe){
-				echo h($maybe['User']['name']).'　';
+				echo $this->Html->link(h($maybe['User']['name']), array('controller' => 'users', 'action' => 'view', $maybe['User']['id']));
+				echo '　';
 			}
 		?>
 	</dd>
@@ -69,17 +77,20 @@
 	<dd>
 		<?php
 			foreach($invited_info as $invited){
-				echo h($invited['User']['name']).'　';
+				echo $this->Html->link(h($invited['User']['name']), array('controller' => 'users', 'action' => 'view', $invited['User']['id']));
+				echo '　';
 			}
 		?>
 	</dd>
- <?php 
-	 if(isset($loginUser)){
-	 	if($event['Event']['user_id'] == $loginUser['id']){
-	 		echo $this->Html->link('Send invitation!', array('action' => 'invite', $event['Event']['id']));
-	 		echo '<br /><br />';
-	 	}
-	 }
+<?php 
+	if(strtotime($event['Event']['open_datetime']) > strtotime(date("Y-m-d H:i:s"))){
+		 if(isset($loginUser)){
+		 	if($event['Event']['user_id'] == $loginUser['id']){
+		 		echo $this->Html->link('Send invitation!', array('action' => 'invite', $event['Event']['id']));
+		 		echo '<br /><br />';
+		 	}
+		 }
+	}
  ?>
 <hr>
 
@@ -88,19 +99,21 @@
 <?php
 	sort($comments);
 	foreach($comments as $comment){
-		echo "　　■".h($comment['Comment']['comment']).'---';
-		echo h($comment['Comment']['created']);
-		echo "(".h($comment['User']['name']).")　　";
-		if(isset($loginUser)){
-			if($comment['Comment']['user_id'] == $loginUser['id']){
-				echo $this->Form->postLink(
-				        'Delete',
-			            array('action' => 'delete_comment', $comment['Comment']['id']),
-			            array('confirm' => 'Are you sure?')
-			            );
+		if($comment['Comment']['del_flg'] != 1){
+			echo "　　■".h($comment['Comment']['comment']).'---';
+			echo h($comment['Comment']['created']);
+			echo "(".h($comment['User']['name']).")　　";
+			if(isset($loginUser)){
+				if($comment['Comment']['user_id'] == $loginUser['id']){
+					echo $this->Form->postLink(
+					        'Delete',
+				            array('action' => 'delete_comment', $comment['Comment']['id']),
+				            array('confirm' => 'Are you sure?')
+				            );
+				}
 			}
+			echo "<br />";
 		}
-		echo "<br />";
 	}
 ?>
